feat: [#186775240] Add health checks for Formation
