+++ api/src/functions/express/app.ts
@@ -281,6 +281,9 @@ const apiFormationClient = ApiFormationClient(
   },
   logger
 );
+
+const apiFormationHealthCheckClient = apiFormationClient.health;
+
 const shouldSaveDocuments = !(process.env.SKIP_SAVE_DOCUMENTS_TO_S3 === "true");
 
 app.use(bodyParser.json({ strict: false }));
@@ -323,6 +326,7 @@ app.use(
       ["dynamics/housing", dynamicsHousingHealthCheckClient],
       ["dynamics/license-status", dynamicsLicenseHealthCheckClient],
       ["webservice/license-status", webServiceLicenseStatusHealthCheckClient],
+      ["webservice/apiFormationHealthCheck", apiFormationHealthCheckClient],
     ])
   )
 );
